attribution/gitminer/requirements.txt
joblib>=0.13.1
GitPython>=2.1.11
pandas>=0.23.4
scipy>=1.2.1
torch>=1.3.1
numpy>=1.14.1
javalang>=0.12.0
tqdm>=4.26.0
